requirements.txt
asgiref==3.7.2
coverage==7.4.4
Django==4.2
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework==3.14.0
exceptiongroup==1.2.0
iniconfig==2.0.0
packaging==24.0
pluggy==1.4.0
psycopg2-binary==2.9.6
pytest==8.1.1
pytz==2024.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
